--- com/bridgelabz/LinkedListMain.java
@@ -1,15 +1,0 @@
-package com.bridgelabz;
-public class LinkedListMain {
-	
-public static void main(String[] args) {
-	
-   	System.out.println("welcome to linked list program");
-			LinkedList list = new LinkedList();
-			list.append(56);
-			list.append(30);
-			list.append(70);
-			list.show();	
-		}
-
-	}
-
